Renamed content-cut.js, content-paste.js, content-copy.js by removing content- to increase consistency with material.io/icons
