+++ src/svg-icons/content/copy.js
