+++ src/svg-icons/content/cut.js
