+++ src/svg-icons/content/paste.js
