+++ src/svg-icons/index.js
@@ -355,9 +355,9 @@ export ContentArchive from './content/archive';
 export ContentBackspace from './content/backspace';
 export ContentBlock from './content/block';
 export ContentClear from './content/clear';
-export ContentContentCopy from './content/content-copy';
-export ContentContentCut from './content/content-cut';
-export ContentContentPaste from './content/content-paste';
+export ContentCopy from './content/copy';
+export ContentCut from './content/cut';
+export ContentPaste from './content/paste';
 export ContentCreate from './content/create';
 export ContentDeleteSweep from './content/delete-sweep';
 export ContentDrafts from './content/drafts';
